refactor: ファイルの分割・ディレクトリ構成の整理
