+++ src/components/config/index.tsx
@@ -3,8 +3,8 @@ import { RiCloseLine } from 'react-icons/ri'
 
 import LogoImg from '../../images/logo.png'
 
-import Section, { ConfigSection } from './section'
-import SelectDir from './selectDir'
+import Footer from './footer'
+import Sections from './sections'
 
 type Props = {
   onClickClose: () => void
@@ -17,41 +17,6 @@ const Config = ({ onClickClose }: Props): JSX.Element => {
     window.api.focusView()
   }
 
-  // ディレクトリを選択
-  const handleClickSelectDir = () => {
-    window.api.showSelectDirDialog()
-  }
-
-  // プライバシーポリシーを開く
-  const handleClickPrivacyPolicy = () => {
-    window.api.openPrivacyPolicy()
-  }
-
-  // 項目ボタン
-  const sections: ConfigSection[] = [
-    {
-      title: 'キャッシュを削除',
-      btnText: '削除',
-      btnBg: 'bg-red-400',
-      btnHoverBg: 'bg-red-600',
-      onClick: () => window.api.removeCache()
-    },
-    {
-      title: '初期化（ログアウト）',
-      btnText: '初期化',
-      btnBg: 'bg-red-400',
-      btnHoverBg: 'bg-red-600',
-      onClick: () => window.api.removeCookie()
-    },
-    {
-      title: '更新を確認',
-      btnText: '確認',
-      btnBg: 'bg-gray-600',
-      btnHoverBg: 'bg-gray-800',
-      onClick: () => window.api.checkUpdate()
-    }
-  ]
-
   return (
     <div className="flex flex-col justify-center items-center min-h-screen bg-neutral-50 drag">
       <button
@@ -62,21 +27,8 @@ const Config = ({ onClickClose }: Props): JSX.Element => {
       </button>
       <div className="w-72 drag-none">
         <img src={LogoImg} alt="serizawa" />
-        <div className="mt-6">
-          <SelectDir onClick={handleClickSelectDir} />
-          {sections.map((e) => (
-            <Section key={e.title} {...e} />
-          ))}
-        </div>
-        <span className="block mt-8 text-center text-xs text-gray-700">
-          {`Developed by arrow2nd - v${process.env.VERSION}`}
-        </span>
-        <span
-          className="block mt-4 text-center text-xs text-gray-700 hover:text-neutral-900 underline cursor-pointer"
-          onClick={handleClickPrivacyPolicy}
-        >
-          プライバシーポリシー
-        </span>
+        <Sections />
+        <Footer />
       </div>
     </div>
   )
